Desktop, Mobile 분기 나눔 & Header, 락락 페이지 반응형으로 수정
