+++ package.json
@@ -16,6 +16,7 @@
     "dayjs": "^1.11.13",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-responsive": "^10.0.1",
     "tailwindcss": "^4.0.17"
   },
   "devDependencies": {
