+++ src/components/Mobile/pages/MobileCongestion.tsx
@@ -1,0 +1,7 @@
+export default function MobileCongestion() {
+    return(
+        <div className="flex justify-center bg-[#F8F4F1]">
+            <img src="../Frame 33.svg" />
+        </div>
+    )
+}
