+++ src/layouts/DesktopLayout.tsx
@@ -1,0 +1,10 @@
+import DesktopHeader from '../components/Desktop/DesktopHeader';
+
+export default function DesktopLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+    <DesktopHeader />
+    {children}
+    </>
+  );
+}
