+++ src/layouts/MobileLayout.tsx
@@ -1,0 +1,10 @@
+import MobileHeader from '../components/Mobile/MobileHeader';
+
+export default function MobileLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+    <MobileHeader />
+    {children}
+    </>
+  );
+}
