+++ src/routes/__root.tsx
@@ -1,13 +1,26 @@
-import { createRootRoute, Outlet } from "@tanstack/react-router";
-import { TanStackRouterDevtools } from "@tanstack/react-router-devtools";
-import Header from "../routes/menu/index";
+import { createRootRoute, Outlet } from '@tanstack/react-router';
+import { useIsMobile, MobileProvider } from '../contexts/MobileContext';
+import MobileLayout from '../layouts/MobileLayout';
+import DesktopLayout from '../layouts/DesktopLayout';
+
+const RootLayout = () => {
+  const { isMobile } = useIsMobile();
+
+  return isMobile ? (
+    <MobileLayout>
+      <Outlet />
+    </MobileLayout>
+  ) : (
+    <DesktopLayout>
+      <Outlet />
+    </DesktopLayout>
+  );
+};
 
 export const Route = createRootRoute({
   component: () => (
-    <>
-      <Header />
-      <Outlet />
-      <TanStackRouterDevtools />
-    </>
+    <MobileProvider>
+      <RootLayout />
+    </MobileProvider>
   ),
 });
