+++ src/routes/menu/congestion.tsx
@@ -1,13 +1,13 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute } from '@tanstack/react-router';
+import { useIsMobile } from '../../contexts/MobileContext';
+import DesktopCongestion from '../../components/Desktop/pages/DesktopCongestion';
+import MobileCongestion from '../../components/Mobile/pages/MobileCongestion';
 
 export const Route = createFileRoute('/menu/congestion')({
-  component: RouteComponent,
-})
+  component : RouteComponent,
+});
 
-function RouteComponent() {
-  return(
-    <div className="flex justify-center bg-[#F8F4F1]">
-        <img src="../Frame 33.svg" />
-    </div>
-  )
+function RouteComponent(){
+    const { isMobile } = useIsMobile();
+    return isMobile ? <MobileCongestion /> : <DesktopCongestion />
 }
